Remove /sitemap.txt, replace with parser.urls
So, the concept of `/sitemap.txt` had a problem - that the
URLs in it were relative (`/docs/something`), which Google
can't understand. It wasn't absolute URLs like
`https://snapcraft.io/docs/something`.

The docs module doesn't currently have any awareness of the domain
of the site, so I think this should be the responsibility of the site
project itself to provide.

So here I remove the `/sitemap.txt` endpoint, and instead add a simple
`urls` list to `parser`, so that this can be used in views in the site
itself.